+++ front/src/components/Inputs/index.ts
@@ -3,3 +3,4 @@ export { InputImage, type InputImageType } from '@/components/Inputs/InputImage'
 export { InputTextArea } from '@/components/Inputs/InputTextArea'
 export { InputLabel } from '@/components/Inputs/InputLabel'
 export { InputErrorMessage } from '@/components/Inputs/InputErrorMessage'
+export { InputSwitch } from '@/components/Inputs/InputSwitch'
