--- front/src/components/Switches/index.ts
@@ -1,1 +1,0 @@
-export { ToggleSwitch } from '@/components/Switches/ToggleSwitch'
